sota_tools: generate test certificates on build
