--- tests/fake_http_server/ca.crt
@@ -1,14 +1,0 @@
------BEGIN CERTIFICATE-----
-MIICMjCCAdigAwIBAgIJAIBVMAWe6TIAMAoGCCqGSM49BAMCMGwxCzAJBgNVBAYT
-AkdCMQ8wDQYDVQQIDAZMb25kb24xDzANBgNVBAcMBkxvbmRvbjEVMBMGA1UECgwM
-VGVzdCBST09UIENBMRAwDgYDVQQLDAdST09UIENBMRIwEAYDVQQDDAlsb2NhbGhv
-c3QwHhcNMTkwNjI0MTI1NjUyWhcNMzkwNjE5MTI1NjUyWjBsMQswCQYDVQQGEwJH
-QjEPMA0GA1UECAwGTG9uZG9uMQ8wDQYDVQQHDAZMb25kb24xFTATBgNVBAoMDFRl
-c3QgUk9PVCBDQTEQMA4GA1UECwwHUk9PVCBDQTESMBAGA1UEAwwJbG9jYWxob3N0
-MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEKC0kxvxrMlZhtqwtW16m2fxXLzSF
-IkvsxffUWH33+2FRXDDz2IBFWkFOISWW7uaINZxAZ/VnMLgJxlHI+Ilgk6NjMGEw
-HQYDVR0OBBYEFJloSlj2VR0GgKNMKs9+4Z+d5xtlMB8GA1UdIwQYMBaAFJloSlj2
-VR0GgKNMKs9+4Z+d5xtlMA8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgGG
-MAoGCCqGSM49BAMCA0gAMEUCIQCHSmp1KEdPXesAZ89mmK87i+fYtqsOFig5O3tP
-iYDjQgIgE+7JkDn2W72YMryIRgsmFRgvdy/jX6KXdKnlXuQC9po=
------END CERTIFICATE-----
