--- tests/sota_tools/cert_generation/tufrepo.url
@@ -1,1 +1,0 @@
-https://localhost:1443/
